requirements.txt
matplotlib
numpy
torch
pygame
pyyaml
# safetensors
# pip install --pre torch --index-url https://download.pytorch.org/whl/nightly/cpu
